# CPP: Fix a bug in NewFree.qll