--- a/cpp/ql/test/query-tests/Critical/NewFree/test.cpp
+++ b/cpp/ql/test/query-tests/Critical/NewFree/test.cpp
@@ -152,7 +152,7 @@ void test3()
 	void *b = my_malloc_3(10);
 	
 	free(a); // GOOD
-	delete b; // BAD: malloc -> delete
+	delete b; // BAD: malloc -> delete [NOT DETECTED]
 }
 
 void test4(bool do_array_delete)
@@ -422,5 +422,5 @@ void test13()
 	MyPointer13 myPointer2(myBuffer);
 	MyPointer13 myPointer3(new char[100]);
 
-	delete myPointer2.getPointer(); // GOOD [FALSE POSITIVE]
+	delete myPointer2.getPointer(); // GOOD
 }