See https://discuss.lgtm.com/t/malloc-delete-mismatch-false-positive/1914.

The issue was that the call to `tmp.data()` was being misrecognized as an allocation call, due to over-enthusiastic data-flow tracking.  The fix is to only use local data flow for the purposes of identifying `malloc`-like functions (whereas global flow is still used to track between the `malloc` and `free`).

Diff query: https://lgtm.com/query/3658696700719411428/