@geoffw0 2019-03-21
> How is performance on the snapshots that recently made this query act up?

I just ran the three queries that use `NewDelete.qll` on wireshark.  The first query goes from 643s -> 809s with these changes.  The other two (with a warmed up cache) stay at around 21s and 12s.  Similar results for one of the projects we had problems with recently - Singular (167s -> 226s on the first query).  And finally 'qtbindings' I don't have and can't get a snapshot of, but qt behaves as expected (379s -> 470s on the first query).

It's likely that with a warmed up cache these queries perform well.  In the long run we might replace the remaining SSA in this library with dataflow, which I'd expect to improve performance.

> Do you think this bug is new in 1.20?

I don't think so, though we have made some changes to the query lately.  It's certainly plausible that the particular FP in the ticket above didn't show up until recently 1.20.