@jbj 2019-03-21
I'm not worried about performance on a cold cache. The data flow library has to be computed at some point in the suite anyway, so the suite won't get slower just because this query uses it. If the second and third queries take 21s and 12s, that means `NewFree.qll` is likely still fast since there's no `cached` in that file.